実数の問題作 成関数を修正、問題出力時に前回の回答など入力内容をリセットするように設定

+++ study_javascript/math_study_room/backend/routes/schema.py
@@ -10,4 +10,4 @@ class QuestionRequestPayload(BaseModel):
 
 class QuestionResponseGeneric(BaseModel):
     questions: list[str]
-    answers: list[int]
+    answers: list[float]
